experimental state refactor
