+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "layer8-interceptor-rs"
-version = "0.1.3"
+version = "0.2.0"
 edition = "2024"
 authors = ["Osoro Bironga <fanosoro@gmail.com>"]
 description = "This repository contains a Rust implementation of the Layer8 Interceptor. Analogous to <https://github.com/globe-and-citizen/layer8-interceptor>"
@@ -43,7 +43,7 @@ js-sys = "0.3"
 serde-wasm-bindgen = "0.6"
 uuid = { version = "1", features = ["js"] }
 tokio = { version = "1", optional = true, features = ["sync"] }
-layer8-primitives = "0.2"
+layer8-primitives = { git = "https://github.com/globe-and-citizen/layer8-primitives-rs.git", branch = "feat/send-status-to-caller" }
 
 [dev-dependencies]
 wasm-bindgen-test = "0.3.50"
