+++ src/lib.rs
@@ -1,10 +1,11 @@
-pub mod health_check;
 pub mod js;
 #[cfg(feature = "websocket")]
 pub mod websocket;
 
 pub(crate) mod js_glue;
-pub(crate) mod types;
+pub(crate) mod network_state;
+mod types;
+
 pub(crate) mod js_imports_prelude {
     use crate::js_glue;
     pub use js_glue::js_imports::{clear_expired_cache, console_error, console_log, object_entries, serve_static};
